% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimedispersion.R
\name{crimedispersion}
\alias{crimedispersion}
\title{Calculate crime dispersion}
\usage{
crimedispersion(data1, unitID, time1, time2, method = "match")
}
\arguments{
\item{data1}{data frame with a minimum of 3 columns with area ID, Rt1 count, Rt2 count}

\item{unitID}{a vector of identifiers for spatial units}

\item{time1}{a vector of crime counts at time 1 (numeric object)}

\item{time2}{a vector of crime counts at time 2 (numeric object)}
}
\value{
A list containing a data frame, values, and a plot
}
\description{
A function to calculate the offence dispersion index (ODI) for crime counts
in unit subregions of an overall region at two time periods, t1 and t2.
}
\references{
Ratcliffe, JH (2010) The spatial dependency of crime increase dispersion, Security Journal, 23(1): 18-36.
}
